Create layouts and posts.
Create an override layout that puts a link to the main page and posts at
the top of each page.  Under that is a table of contents for the page.

Create an override layout for pages that takes content from the layout
above and adds the post date, author, and title to a post.  It also
allows for an update date.

Create a posts page to display all the posts with the date they were
posted.

Move contents of "index.md" into "_posts/2021-03-09-getting_started.md",
and reference it as such from the main page.